LayoutTests/imported/w3c/web-platform-tests/css/css-overscroll-behavior/overscroll-behavior-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: action

overscroll-behavior

Tests that overscroll-behavior prevents scroll-propagation in the area and direction as specified.

Make two scrolls on BLUE, in this order: scroll UP (or drag down), then scroll LEFT (or drag right). Scroll (or drag) until nothing is scrolling. Then tap on DONE.
Repeat the same scrolls as in step 1 and then tap on DONE.
Repeat the same scrolls as in step 1 and then tap on DONE.
Make two separate scrolls on GREEN, in this order: scroll UP (or drag down), then scroll LEFT (or drag right). Scroll (or drag) until nothing is scrolling. Then tap on DONE.

Harness Error (FAIL), message = Unhandled rejection: Can't find variable: action

NOTRUN overscroll-behavior prevents scroll-propagation in the area and direction as specified
